server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Mohamed <mohamed.sheik98@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
autopep8 = "^2.0.0"
mypy = "^0.990"
django = "^4.1.3"
djangorestframework = "^3.14.0"
cors = "^1.0.1"
django-cors-headers = "^3.13.0"
django-address = "^0.2.8"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
